requirements.txt
streamlit
langchain
langchain-google-genai
langchain-community
faiss-cpu
youtube-transcript-api
python-dotenv
google-generativeai